nb/amd/amdfam10: Use 64 bits in multiplication
The literal needs to be cast to a uint64 to prevent num_nodes
from being implicitly promoted to a signed int.

Found-by: Coverity CID 1347343
Reviewed-on: https://review.coreboot.org/c/coreboot/+/33210
Tested-by: build bot (Jenkins) <no-reply@coreboot.org>

+++ src/northbridge/amd/amdfam10/northbridge.c
@@ -779,7 +779,7 @@ static void amdfam10_domain_read_resources(struct device *dev)
 
 			/* Calculate CC6 storage area size */
 			if (interleaved)
-				qword = (0x1000000 * num_nodes);
+				qword = (uint64_t)0x1000000 * num_nodes;
 			else
 				qword = 0x1000000;
